pyproject.toml
[tool.poetry]
name = "weather-data"
version = "0.1.0"
description = ""
authors = ["Bilel <debbabi.bilel@gmail.com>"]
readme = "README.md"
packages = [{include = "weather-data"}]

[tool.poetry.dependencies]
python = "^3.10"
paho-mqtt = "^1.6.1"
python-dotenv = "^1.0.0"
geocoder = "^1.38.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
